requirements.txt
anthropic==0.39.0
httpx==0.28.1
apscheduler==3.10.4
python-dotenv==1.0.1
aiosqlite==0.20.0
pyyaml==6.0.2
